requirements.txt
xlrd
# TODO: add specific versions of the packages
sentence-transformers
# see https://pytorch.org/get-started/locally/
torch
langchain_community
langchain
langchain_core
pandas
pandera
streamlit
scikit-learn
umap-learn